fix(auth): Prevent redirect to login after reconnection
Two issues were fixed. One was a premature redirect to the login page after a user reconnects to the app. This was due to an improper logic in 'refresh_token.go' and 'root.go' that redirects to '/account/login' after successful authentication

The second issue was an incorrect error render logic in 'account.go'. It will now properly show the error message on invalid credentials

Also, 'hx-swap' is set to none to prevent element swapping after a GET request to '/api/token/refresh'